Fixed custom server issue & adjusted rpcs

+++ src/Gamemodes/Colorwars/ColorwarsAssignRoles.cs
@@ -47,8 +47,8 @@ public class ColorwarsAssignRoles
             foreach (PlayerControl player in team)
             {
                 if (player.PlayerId == localPlayer.PlayerId) player.SetRole(RoleTypes.Impostor);
-                RpcV2.Immediate(player.NetId, (byte)RpcCalls.SetRole).Write((byte)RoleTypes.Impostor).SendToFollowing(teamClientIds);
-                RpcV2.Immediate(player.NetId, (byte)RpcCalls.SetRole).Write((byte)RoleTypes.Crewmate).SendToAll(teamClientIds);
+                RpcV2.Immediate(player.NetId, (byte)RpcCalls.SetRole).Write((ushort)RoleTypes.Impostor).SendToFollowing(teamClientIds);
+                RpcV2.Immediate(player.NetId, (byte)RpcCalls.SetRole).Write((ushort)RoleTypes.Crewmate).SendToAll(teamClientIds);
                 Game.AssignRole(player, CustomRoleManager.Static.SerialKiller);
             }
         }
